+++ src/playground/src/components/editor.tsx
@@ -1,23 +1,29 @@
-import loader from '@monaco-editor/loader';
-import {useEffect, useRef} from "react";
-import {editor} from "monaco-editor";
-import IStandaloneEditorConstructionOptions = editor.IStandaloneEditorConstructionOptions;
+import {FC, useEffect, useState} from "react";
+import {CadencePlayer} from "@/components/player.tsx";
+import {SampleDirectory} from "@/samplesTypes.ts";
+import {Editor, Monaco} from "@monaco-editor/react";
+import {sampleDirectories} from "@/assets-list.ts";
+
 
 const baseValue = `
-import A3Piano from "./assets/samples/piano/A3.mp3"
-const loop: Loop = {
-    sample: A3Piano,
-    interval: "1s",
-    speed: 1
-}
+import type {Loop} from "cadence-js";
+import {Cadence} from "cadence-js";
+import {E2} from "guitar-nylon";
 
-export default loop
+const cadence = new Cadence();
+const loop: Loop = {
+    interval: "1s",
+    speed: 1,
+    sample: E2
+}
+cadence.play(loop)
 `
 
-async function fetchCadenceTypes (): Promise<string> {
+
+async function fetchCadenceTypes(): Promise<string> {
     try {
         const types = await fetch('cadence.d.ts').then((res) => res.text());
-        return types
+        return types.replace("declare module \"index\"", 'declare module "cadence-js"')
     } catch (e) {
         console.error(e);
         return ''
@@ -25,31 +31,56 @@ async function fetchCadenceTypes (): Promise<string> {
 
 }
 
-async function createEditor(wrapper: HTMLDivElement) {
-    const types = await fetchCadenceTypes();
-    const monaco = await loader.init();
-    const libUri = 'ts:filename/cadence.d.ts';
-    monaco.languages.typescript.typescriptDefaults.addExtraLib(types, libUri);
-    console.log("set cadence types", types)
+function makeDeclarationFile(sampleDirectories: SampleDirectory[]): string {
+    let declarationFile = '';
 
-    const properties : IStandaloneEditorConstructionOptions = {
-        value:baseValue,
-        language: 'typescript',
-        theme: 'vs-dark',
-    };
+    sampleDirectories.forEach((directory) => {
+        declarationFile += `declare module "${directory.name}" {\n`;
+        directory.samples.forEach((sample) => {
+            declarationFile += `export const ${sample.name}: Loop;\n`;
+        });
+        declarationFile += `}\n`;
+    })
 
-    monaco.editor.create(wrapper, properties);
+
+    return declarationFile
 }
 
-export function Editor() {
-    const elementRef = useRef<HTMLDivElement >(null);
+async function createEditor(monaco: Monaco) {
+    const types = await fetchCadenceTypes();
+    const samplesDeclaration = makeDeclarationFile(sampleDirectories);
+    monaco.languages.typescript.typescriptDefaults.addExtraLib(types, "cadence-js");
+    monaco.languages.typescript.typescriptDefaults.addExtraLib(samplesDeclaration, "samples");
+    monaco.languages.typescript.typescriptDefaults.setCompilerOptions({
+        allowNonTsExtensions: true,
+        noLib: true
+    });
+}
+
+
+export const CadenceEditor: FC<{mustPlay:boolean}> = ({mustPlay}) => {
+    const [code, setCode] = useState<string>('')
+
     useEffect(() => {
-        if (elementRef.current) {
-            createEditor(elementRef.current);
-        }
+        setCode(baseValue)
     }, []);
 
+
     return (
-        <div className={"min-h-screen"} ref={elementRef}></div>
+        <div className={"min-h-screen"}>
+            <Editor
+                width="500px"
+                height="500px"
+                theme="vs-dark"
+                defaultLanguage="typescript"
+                beforeMount={createEditor}
+                defaultValue={code}
+                onChange={(value) => {
+                    setCode(value ?? '')
+                }
+                }
+            />;
+            <CadencePlayer cadenceCode={code} mustPlay={mustPlay}/>
+        </div>
     )
 }
